Folders structure refactor
