+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "bananacss",
-  "description": "The brazilian css superset",
-  "version": "0.5.0",
+  "description": "The Brazilian css superset",
+  "version": "0.5.1",
   "author": {
     "name": "afonsopacifer",
     "email": "afonsopacifer@live.com"
@@ -37,7 +37,7 @@
     "test": "mocha test/index.js",
     "cover": "istanbul cover _mocha -- ./test/* --recursive",
     "coveralls": "npm run cover -- --report lcovonly && cat ./coverage/lcov.info | coveralls",
-    "eslint": "eslint src/* bin/* test/*",
+    "eslint": "eslint src/* src/core/* src/helpers/* bin/* test/*",
     "jsdocs": "jsdoc bin/*.js src/*.js",
     "publish jsdocs": "git push origin `git subtree split --prefix out master`:gh-pages --force"
   },
