+++ src/core/bnnHeight.js
@@ -1,10 +1,10 @@
-const getParam = require('./getParam.js');
-const addProperty = require('./addProperty.js');
-const removeProperty = require('./removeProperty.js');
+const getParam = require('../helpers/getParam.js');
+const addProperty = require('../helpers/addProperty.js');
+const removeProperty = require('../helpers/removeProperty.js');
 
 /**
  * Compile the bnn-height property values into height and max-height.
- * @module src/bnnHeight
+ * @module src/core/bnnHeight
  * @param {array} declarations - Declarations list for a single CSS rule (AST)
  */
 const bnnHeight = (declarations) => {
