+++ src/core/bnnPosition.js
@@ -1,11 +1,11 @@
-const getParam = require('./getParam.js');
-const addProperty = require('./addProperty.js');
-const removeProperty = require('./removeProperty.js');
+const getParam = require('../helpers/getParam.js');
+const addProperty = require('../helpers/addProperty.js');
+const removeProperty = require('../helpers/removeProperty.js');
 
 /**
  * Compile the bnn-position property values
  * into top, right, left and bottom or centralize with margins.
- * @module src/bnnPosition
+ * @module src/core/bnnPosition
  * @param {array} declarations - Declarations list for a single CSS rule (AST)
  */
 const bnnPosition = (declarations) => {
