+++ src/helpers/getParam.js
@@ -1,6 +1,6 @@
 /**
  * Get a especific value from a shorthand
- * @module src/getParam
+ * @module src/helpers/getParam
  * @param {string} declarationValue - Shorthand value of a declaration
  * @param {number} paramPosition - Position for returned value
  */
